# backend_services/gateway/docs/swagger.yaml
basePath: /
definitions:
...
  internal_handler.AppGetResponse:
    properties:
      app:
        $ref: '#/definitions/github_com_eternal-orbit-labs_gateway_internal_domain.App'
    type: object
  internal_handler.AppListResponse:
    properties:
      apps:
        items:
          $ref: '#/definitions/github_com_eternal-orbit-labs_gateway_internal_domain.App'
        type: array
      next_cursor:
        example: ""
        type: string
    type: object
...
  internal_handler.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  internal_handler.OrgListResponse:
    properties:
      invites:
        items:
          $ref: '#/definitions/github_com_eternal-orbit-labs_gateway_internal_domain.AppOrgInviteListItem'
        type: array
      organizations:
        items:
          $ref: '#/definitions/github_com_eternal-orbit-labs_gateway_internal_domain.AppOrgListItem'
        type: array
    type: object
  internal_handler.OrgResponse:
    properties:
      organization:
        $ref: '#/definitions/github_com_eternal-orbit-labs_gateway_internal_domain.AppOrgListItem'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API gateway for Eternal Orbit Labs — authentication, organizations,
    teams, and app management.
  title: EOL Gateway API
  version: "1.0"
paths:
  /api/apps:
    get:
      description: Returns a paginated list of apps. Supports cursor-based pagination
        and search.
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - default: 20
        description: Results per page (max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AppListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: List apps
      tags:
      - Apps
  /api/apps/{slug}:
    get:
      description: Returns a single app by its URL slug.
      parameters:
      - description: App slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AppGetResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Get app by slug
      tags:
      - Apps
  /api/apps/{slug}/invites/{inviteId}/accept:
    post:
      description: Accept a pending organization invite. The user must be the invite
        recipient.
      parameters:
      - description: App slug
        in: path
        name: slug
        required: true
        type: string
      - description: Invite ID
        in: path
        name: inviteId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.OrgResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Accept invite
      tags:
      - Invites
  /api/apps/{slug}/orgs:
    get:
      description: Returns the authenticated user's organizations and pending invites
        for the given app.
      parameters:
      - description: App slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.OrgListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: List user's organizations
      tags:
      - Organizations
    post:
      consumes:
      - application/json
      description: Create a new organization under the given app. The authenticated
        user becomes the owner.
      parameters:
      - description: App slug
        in: path
        name: slug
        required: true
        type: string
      - description: Organization details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateOrgRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.OrgResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Create organization
      tags:
      - Organizations
  /api/apps/{slug}/orgs/{orgId}:
    get:
      description: Returns organization details. Requires the user to be a member.
      parameters:
      - description: App slug
        in: path
        name: slug
        required: true
        type: string
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.OrgResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Get organization
      tags:
      - Organizations
  /api/apps/{slug}/orgs/{orgId}/invites:
    post:
      consumes:
      - application/json
      description: Send an invitation to join the organization.
      parameters:
      - description: App slug
        in: path
        name: slug
        required: true
        type: string
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Invite details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateInviteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.InviteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Create invite
      tags:
      - Invites
  /api/apps/{slug}/orgs/{orgId}/invites/{inviteId}:
    delete:
      description: Permanently delete a pending invitation.
      parameters:
      - description: App slug
        in: path
        name: slug
        required: true
        type: string
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Invite ID
        in: path
        name: inviteId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Delete invite
      tags:
      - Invites
    patch:
      consumes:
      - application/json
      description: Update a pending invite's details (email, name, phone, status).
      parameters:
      - description: App slug
        in: path
        name: slug
        required: true
        type: string
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Invite ID
        in: path
        name: inviteId
        required: true
        type: string
      - description: Updated invite details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateInviteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.InviteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Update invite
      tags:
      - Invites
  /api/apps/{slug}/orgs/{orgId}/members:
    get:
      description: Returns a paginated list of organization members with optional
        filters.
      parameters:
      - description: App slug
        in: path
        name: slug
        required: true
        type: string
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Results per page (max 100)
        in: query
        name: limit
        type: integer
      - description: Search across name, email, phone
        in: query
        name: q
        type: string
      - description: Filter by first name
        in: query
        name: first_name
        type: string
      - description: Filter by last name
        in: query
        name: last_name
        type: string
      - description: Filter by email
        in: query
        name: email
        type: string
      - description: Filter by phone
        in: query
        name: phone
        type: string
      - description: Filter by status (active, deactive)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MemberListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: List org members
      tags:
      - Members
    post:
      consumes:
      - application/json
      description: Add an existing user to the organization by email.
      parameters:
      - description: App slug
        in: path
        name: slug
        required: true
        type: string
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Member details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateMemberRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Add member
      tags:
      - Members
  /api/apps/{slug}/orgs/{orgId}/members/{memberId}:
    delete:
      description: Remove a member from the organization. Cannot remove the owner.
      parameters:
      - description: App slug
        in: path
        name: slug
        required: true
        type: string
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Member user ID
        in: path
        name: memberId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Remove member
      tags:
      - Members
    patch:
      consumes:
      - application/json
      description: Activate or deactivate an organization member.
      parameters:
      - description: App slug
        in: path
        name: slug
        required: true
        type: string
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Member user ID
        in: path
        name: memberId
        required: true
        type: string
      - description: New status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateMemberRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Update member status
      tags:
      - Members
  /api/apps/{slug}/orgs/{orgId}/sidebar:
    get:
      description: Returns the sidebar navigation configuration for an organization
        workspace.
      parameters:
      - description: App slug
        in: path
        name: slug
        required: true
        type: string
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SidebarResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Get org sidebar
      tags:
      - Sidebar
  /api/apps/{slug}/orgs/{orgId}/teams:
    get:
      description: Returns a paginated list of teams in the organization.
      parameters:
      - description: App slug
        in: path
        name: slug
        required: true
        type: string
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Results per page (max 100)
        in: query
        name: limit
        type: integer
      - description: Search by team name
        in: query
        name: q
        type: string
      - description: Filter by team name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.TeamListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: List teams
      tags:
      - Teams
    post:
      consumes:
      - application/json
      description: Create a new team in the organization.
      parameters:
      - description: App slug
        in: path
        name: slug
        required: true
        type: string
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Team details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateTeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.TeamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Create team
      tags:
      - Teams
  /api/apps/{slug}/orgs/{orgId}/teams/{teamId}:
    get:
      description: Returns team details by ID.
      parameters:
      - description: App slug
        in: path
        name: slug
        required: true
        type: string
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Team ID
        in: path
        name: teamId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.TeamResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Get team
      tags:
      - Teams
...
  /api/apps/{slug}/orgs/{orgId}/teams/{teamId}/members/{userId}:
    delete:
      description: Remove a single member from the team.
      parameters:
      - description: App slug
        in: path
        name: slug
        required: true
        type: string
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Team ID
        in: path
        name: teamId
        required: true
        type: string
      - description: User ID to remove
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Remove team member
      tags:
      - Teams
...
  /api/auth/apikeys/{id}:
    delete:
      description: Permanently delete a personal API key by ID.
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Delete API key
      tags:
      - API Keys
...
  /api/auth/google:
    get:
      description: Redirects the user to Google's consent screen. Sets an oauth_state
        cookie for CSRF protection.
      responses:
        "307":
          description: Redirect to Google
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Start Google OAuth
      tags:
      - OAuth
...
schemes:
- http
- https
securityDefinitions:
  BearerAPIKey:
    description: Personal API key — send as "Bearer eol_k1_..."
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    description: JWT access token set automatically on login/signup.
    in: cookie
    name: eol_access
    type: apiKey
swagger: "2.0"